+++ src/bucket/BucketTests.cpp
@@ -43,27 +43,27 @@ mask(uint32_t v, uint32_t m)
     return (v & ~(m - 1));
 }
 uint32_t
-size(size_t level)
+size(uint32_t level)
 {
     return 1 << (2 * (level + 1));
 }
 uint32_t
-half(size_t level)
+half(uint32_t level)
 {
     return size(level) >> 1;
 }
 uint32_t
-prev(size_t level)
+prev(uint32_t level)
 {
     return size(level - 1);
 }
 uint32_t
-lowBoundExclusive(size_t level, uint32_t ledger)
+lowBoundExclusive(uint32_t level, uint32_t ledger)
 {
     return mask(ledger, size(level));
 }
 uint32_t
-highBoundInclusive(size_t level, uint32_t ledger)
+highBoundInclusive(uint32_t level, uint32_t ledger)
 {
     return mask(ledger, prev(level));
 }
@@ -84,8 +84,8 @@ countEntries(std::shared_ptr<Bucket> bucket)
 }
 
 void
-checkBucketSizeAndBounds(BucketList& bl, uint32_t ledgerSeq,
-                         uint32_t level, bool isCurr)
+checkBucketSizeAndBounds(BucketList& bl, uint32_t ledgerSeq, uint32_t level,
+                         bool isCurr)
 {
     std::shared_ptr<Bucket> bucket;
     uint32_t sizeOfBucket = 0;
@@ -125,12 +125,10 @@ checkBucketSizeAndBounds(BucketList& bl, uint32_t ledgerSeq,
 // If pred is false for ledger < L and true for ledger >= L then
 // binarySearchForLedger will return L.
 uint32_t
-binarySearchForLedger(
-        uint32_t lbound,
-        uint32_t ubound,
-        const std::function<uint32_t(uint32_t)>& pred)
+binarySearchForLedger(uint32_t lbound, uint32_t ubound,
+                      const std::function<uint32_t(uint32_t)>& pred)
 {
-    while (lbound+1 != ubound)
+    while (lbound + 1 != ubound)
     {
         uint32_t current = (lbound + ubound) / 2;
         if (pred(current))
@@ -264,7 +262,7 @@ TEST_CASE("bucket list", "[bucket]")
             if (i % 10 == 0)
                 CLOG(DEBUG, "Bucket") << "Added batch " << i
                                       << ", hash=" << binToHex(bl.getHash());
-            for (size_t j = 0; j < BucketList::kNumLevels; ++j)
+            for (uint32_t j = 0; j < BucketList::kNumLevels; ++j)
             {
                 auto const& lev = bl.getLevel(j);
                 auto currSz = countEntries(lev.getCurr());
@@ -324,7 +322,7 @@ TEST_CASE("bucket list shadowing", "[bucket]")
             CLOG(DEBUG, "Bucket") << "Added batch " << i
                                   << ", hash=" << binToHex(bl.getHash());
             // Alice and bob should be in either curr or snap of level 0 and 1
-            for (size_t j = 0; j < 2; ++j)
+            for (uint32_t j = 0; j < 2; ++j)
             {
                 auto const& lev = bl.getLevel(j);
                 auto curr = lev.getCurr();
@@ -340,7 +338,7 @@ TEST_CASE("bucket list shadowing", "[bucket]")
 
             // Alice and Bob should never occur in level 2 .. N because they
             // were shadowed in level 0 continuously.
-            for (size_t j = 2; j < BucketList::kNumLevels; ++j)
+            for (uint32_t j = 2; j < BucketList::kNumLevels; ++j)
             {
                 auto const& lev = bl.getLevel(j);
                 auto curr = lev.getCurr();
@@ -384,7 +382,7 @@ TEST_CASE("duplicate bucket entries", "[bucket]")
                 CLOG(DEBUG, "Bucket") << "Added batch " << i
                                       << ", hash1=" << hexAbbrev(bl1.getHash())
                                       << ", hash2=" << hexAbbrev(bl2.getHash());
-            for (size_t j = 0; j < BucketList::kNumLevels; ++j)
+            for (uint32_t j = 0; j < BucketList::kNumLevels; ++j)
             {
                 auto const& lev1 = bl1.getLevel(j);
                 auto const& lev2 = bl2.getLevel(j);
@@ -477,7 +475,7 @@ TEST_CASE("file-backed buckets", "[bucket][bucketbench]")
     CLOG(DEBUG, "Bucket") << "Hashing entries";
     std::shared_ptr<Bucket> b1 =
         Bucket::fresh(app->getBucketManager(), live, dead);
-    for (size_t i = 0; i < 5; ++i)
+    for (uint32_t i = 0; i < 5; ++i)
     {
         CLOG(DEBUG, "Bucket") << "Merging 10000 new ledger entries into "
                               << (i * 10000) << " entry bucket";
@@ -603,7 +601,7 @@ clearFutures(Application::pointer app, BucketList& bl)
 {
 
     // First go through the BL and mop up all the FutureBuckets.
-    for (size_t i = 0; i < BucketList::kNumLevels; ++i)
+    for (uint32_t i = 0; i < BucketList::kNumLevels; ++i)
     {
         bl.getLevel(i).getNext().clear();
     }
@@ -731,10 +729,10 @@ TEST_CASE("single entry bubbling up", "[bucket][bucketbubble]")
 
             CLOG(DEBUG, "Bucket") << "------- ledger " << i;
 
-            for (size_t j = 0; j <= BucketList::kNumLevels - 1; ++j)
+            for (uint32_t j = 0; j <= BucketList::kNumLevels - 1; ++j)
             {
-                size_t lb = lowBoundExclusive(j, i);
-                size_t hb = highBoundInclusive(j, i);
+                uint32_t lb = lowBoundExclusive(j, i);
+                uint32_t hb = highBoundInclusive(j, i);
 
                 auto const& lev = bl.getLevel(j);
                 auto currSz = countEntries(lev.getCurr());
@@ -792,7 +790,7 @@ TEST_CASE("bucket persistence over app restart", "[bucket][bucketpersist]")
     cfg1.ARTIFICIALLY_PESSIMIZE_MERGES_FOR_TESTING = true;
 
     std::vector<std::vector<LedgerEntry>> batches;
-    for (size_t i = 0; i < 110; ++i)
+    for (uint32_t i = 0; i < 110; ++i)
     {
         batches.push_back(LedgerTestUtils::generateValidLedgerEntries(1));
     }
@@ -910,18 +908,21 @@ TEST_CASE("BucketList sizeOf* and oldestLedgerIn* relations", "[bucket][count]")
             if (BucketList::sizeOfSnap(ledger, level) > 0)
             {
                 uint32_t oldestInCurr =
-                        BucketList::oldestLedgerInSnap(ledger, level) +
-                        BucketList::sizeOfSnap(ledger, level);
+                    BucketList::oldestLedgerInSnap(ledger, level) +
+                    BucketList::sizeOfSnap(ledger, level);
                 REQUIRE(oldestInCurr ==
                         BucketList::oldestLedgerInCurr(ledger, level));
             }
             if (BucketList::sizeOfCurr(ledger, level) > 0)
             {
                 uint32_t newestInCurr =
-                        BucketList::oldestLedgerInCurr(ledger, level) +
-                        BucketList::sizeOfCurr(ledger, level) - 1;
-                REQUIRE(newestInCurr == (level == 0 ? ledger :
-                        BucketList::oldestLedgerInSnap(ledger, level-1) - 1));
+                    BucketList::oldestLedgerInCurr(ledger, level) +
+                    BucketList::sizeOfCurr(ledger, level) - 1;
+                REQUIRE(newestInCurr == (level == 0
+                                             ? ledger
+                                             : BucketList::oldestLedgerInSnap(
+                                                   ledger, level - 1) -
+                                                   1));
             }
         }
     }
@@ -932,24 +933,22 @@ TEST_CASE("BucketList snap reaches steady state", "[bucket][count]")
     std::default_random_engine gen;
     // Deliberately exclude deepest level since snap on the deepest level
     // is always empty.
-    for (uint32_t level = 0; level < BucketList::kNumLevels-1; ++level)
+    for (uint32_t level = 0; level < BucketList::kNumLevels - 1; ++level)
     {
         uint32_t const half = BucketList::levelHalf(level);
 
         // Use binary search (assuming that it does reach steady state)
         // to find the ledger where the snap at this level first reaches
         // max size.
-        uint32_t boundary =
-            binarySearchForLedger(1, std::numeric_limits<uint32_t>::max() / 2,
-                    [level, half] (uint32_t ledger)
-                    {
-                        return (BucketList::sizeOfSnap(ledger, level) ==
-                                half);
-                    });
+        uint32_t boundary = binarySearchForLedger(
+            1, std::numeric_limits<uint32_t>::max() / 2,
+            [level, half](uint32_t ledger) {
+                return (BucketList::sizeOfSnap(ledger, level) == half);
+            });
 
         // Generate random ledgers above and below the split to test that
         // it was actually at steady state.
-        std::uniform_int_distribution<uint32_t> distLow(1, boundary-1);
+        std::uniform_int_distribution<uint32_t> distLow(1, boundary - 1);
         std::uniform_int_distribution<uint32_t> distHigh(boundary);
         for (uint32_t i = 0; i < 1000; ++i)
         {
@@ -967,13 +966,12 @@ TEST_CASE("BucketList deepest curr accumulates", "[bucket][count]")
     uint32_t const deepest = BucketList::kNumLevels - 1;
     // Use binary search to find the first ledger where the deepest curr
     // first is non-empty.
-    uint32_t boundary =
-        binarySearchForLedger(1, std::numeric_limits<uint32_t>::max() / 2,
-                [deepest] (uint32_t ledger)
-                {
-                    return (BucketList::sizeOfCurr(ledger, deepest) > 0);
-                });
-    std::uniform_int_distribution<uint32_t> distLow(1, boundary-1);
+    uint32_t boundary = binarySearchForLedger(
+        1, std::numeric_limits<uint32_t>::max() / 2,
+        [deepest](uint32_t ledger) {
+            return (BucketList::sizeOfCurr(ledger, deepest) > 0);
+        });
+    std::uniform_int_distribution<uint32_t> distLow(1, boundary - 1);
     std::uniform_int_distribution<uint32_t> distHigh(boundary);
     for (uint32_t i = 0; i < 1000; ++i)
     {
